Update the location tracking interface, provide a sample LocationManager implementation with the iOS package.

+++ Com.Kumulos.Abstractions/IKumulos.cs
@@ -8,6 +8,7 @@ namespace Com.Kumulos.Abstractions
         void Initialize(IKSConfig config);
 
         string InstallId { get; }
+
         Build Build { get; }
         PushChannels PushChannels { get; }
 
@@ -17,11 +18,15 @@ namespace Com.Kumulos.Abstractions
 
         void TrackEvent(string eventType, Dictionary<string, string> properties);
         void TrackEventImmediately(string eventType, Dictionary<string, string> properties);
+
         void LogException(Exception e);
         void LogUncaughtException(Exception e);
-        void SendLocationUpdate(decimal lat, decimal lng);
+
+        void SendLocationUpdate(double lat, double lng);
+
         void AssociateUserWithInstall(string userIdentifier);
         void AssociateUserWithInstall(string userIdentifier, Dictionary<string, string> attributes);
+
         void TrackEddystoneBeaconProximity(string namespaceHex, string instanceHex, int? distanceMetres);
         void TrackiBeaconProximity(object CLBeaconObject);
     }
